♻️  Refactor app structure: Some folder & file names have been updated to more descriptive names.
