+++ app/notification/magic_link/map_contents.py
@@ -1,6 +1,4 @@
-from app.notification.notification_operations.custom_exceptions import (
-    NotificationError,
-)
+from app.notification.model.exceptions import NotificationError
 from tests.test_magic_link.magic_link_data import (
     expected_magic_link_data,
 )
