+++ app/notification/model/exceptions.py
